+++ src/VaultSharp/V1/SecretsEngines/GoogleCloudKMS/DecryptionResponse.cs
@@ -1,0 +1,16 @@
+﻿using Newtonsoft.Json;
+
+namespace VaultSharp.V1.SecretsEngines.GoogleCloudKMS
+{
+    /// <summary>
+    /// Decryption output.
+    /// </summary>
+    public class DecryptionResponse
+    {
+        /// <summary>
+        ///  Decrypted plain text.
+        /// </summary>
+        [JsonProperty("plaintext")]
+        public string PlainText { get; set; }
+    }
+}