+++ src/VaultSharp/V1/SecretsEngines/GoogleCloudKMS/SignatureResponse.cs
@@ -1,0 +1,16 @@
+﻿using Newtonsoft.Json;
+
+namespace VaultSharp.V1.SecretsEngines.GoogleCloudKMS
+{
+    /// <summary>
+    /// Signature output.
+    /// </summary>
+    public class SignatureResponse
+    {
+        /// <summary>
+        /// The signature
+        /// </summary>
+        [JsonProperty("signature")]
+        public string Signature { get; set; }
+    }
+}