+++ src/VaultSharp/V1/SecretsEngines/GoogleCloudKMS/VerificationResponse.cs
@@ -1,0 +1,16 @@
+﻿using Newtonsoft.Json;
+
+namespace VaultSharp.V1.SecretsEngines.GoogleCloudKMS
+{
+    /// <summary>
+    /// Verification output.
+    /// </summary>
+    public class VerificationResponse
+    {
+        /// <summary>
+        /// Flag to indicate if signature is valid.
+        /// </summary>
+        [JsonProperty("valid")]
+        public bool Valid { get; set; }
+    }
+}