+++ src/VaultSharp/V1/SecretsEngines/SecretsEngineDefaultPaths.cs
@@ -12,6 +12,7 @@
         public const string Cubbyhole = "cubbyhole";
         public const string Database = "database";
         public const string GoogleCloud = "gcp";
+        public const string GoogleCloudKMS = "gcpkms";
         public const string KeyValueV1 = "kv";
         public const string KeyValueV2 = "secret";
         public const string Identity = "identity";